Update Vector_Similarity.py
As an author of this algorithm, I am happy to have a contribution too. Arc Cosine accepts only values between 1 and -1. In this algorithm, Arc Cosine gets the Cosine value.  In some situations, Cosine value is 1.00000000000001 which is out of boundary for Arc Cosine. This is a Python problem in dealing with float numbers properly. Hence  we round the input to Arc Cosine up to 4 decimal points. The issue has been raised before in Stackoverflow as well: https://stackoverflow.com/questions/13637400/why-do-i-get-valueerror-math-domain-error/13637532